fixed bugs and reorganized css part
