+++ src/components/Home.js
@@ -1,18 +1,23 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import Navbar from './common/Navbar'
 import Mytrainings from './MyTrainings/Mytrainings'
 import Trainers from './Trainers/Trainers';
-import { Route,Routes } from 'react-router-dom';
+import { Route,Routes,useNavigate } from 'react-router-dom';
 import About from './About/About';
 import Login from './Login/Login';
 
 
 const Home = () => {
   const [login,setLogin]=useState(true);
+  const navigate = useNavigate();
 
   const handleLogout=(value)=>{
     setLogin(value);
   }
+
+  useEffect(()=>{
+    login?navigate("/",true):navigate("/mytrainings",true);
+  },[login])
 
   return (
     <>
@@ -21,7 +26,7 @@ const Home = () => {
     (<div>
         <Navbar handleLogout={handleLogout}/>
         <Routes>
-          <Route path="/" element={<Mytrainings />}/>
+          <Route path="/mytrainings" element={<Mytrainings />}/>
           <Route path="/trainers" element={<Trainers />}/>
           <Route path="/aboutus" element={<About />}/>
         </Routes>
