+++ src/components/common/Navbar.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import website_logo from "../../images/website_logo.png";
-import "../../styles/common/navbar.css"
+import "./navbar.css"
 import { NavLink } from 'react-router-dom';
 import profile from "../../images/profile.svg"
 import Profile from './Profile';
@@ -14,7 +14,7 @@ const Navbar = (props) => {
     <nav className='navbar'>
         <div className='website_logo_wrapper'><img className='website_logo' src={website_logo}/></div>
         <div className='navlinksContainer'>
-            <NavLink className="navlinks" to={"/"}>MyTrainings</NavLink>
+            <NavLink className="navlinks" to={"/mytrainings"} >MyTrainings</NavLink>
             <NavLink className="navlinks" to={"/trainers"}>Trainers</NavLink>
             <NavLink className="navlinks" to={"/aboutus"}>AboutUs</NavLink>
             <div onClick={()=>{profileflag ? setProfileflag(false):setProfileflag(true)}} className="profile_div">
