--- src/styles/mytrainings/mytrainings.css
@@ -1,46 +1,0 @@
-
-.mytrainingsContainer{
-    /* padding: 20px; */
-}
-
-.mytrainingsContainer>h1{
-    text-align: center;
-    font-size: 36px;
-    padding: 10px 20px;
-    margin: 0px 40px;
-    border-bottom: 1px solid rgba(58, 56, 56, 0.212);
-}
-
-.mytrainings{
-    /* margin: 30px 40px; */
-    display: grid;
-    grid-template-columns: repeat(3,1fr);
-}
-
-
-.mytrainings>div{
-    /* text-align: center; */
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 25vw;
-    height: 25vh;
-    margin: 3vw;
-    cursor: pointer;
-    font-family: Arial, Helvetica, sans-serif;
-    font-weight: bold;
-    font-size: 20px;
-    border-radius: 3px;
-    box-shadow: rgba(0, 0, 0, 0.1) 0px 10px 50px;
-    
-}
-
-
-
-.mytrainings>div:hover{
-    transform: scale(1.02);
-}
-
-.add_icon{
-    font-size: 30px;
-}
